💫 Move code into controller
